[Silver V] Title: 공포의 면담실, Time: 548 ms, Memory: 93728 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/30088. 공포의 면담실/공포의 면담실.java
@@ -17,7 +17,7 @@ public class Main {
             for (int i = 0; i < n; i++) sum += Integer.parseInt(st.nextToken());
             arr[tc] = sum;
         }
-        int answer = 0;
+        long answer = 0;
         Arrays.sort(arr);
         for (int i = 1; i < t + 1; i++) {
             arr[i] = arr[i] + arr[i - 1];
